Add lazy loading component
Wordpress application renders a react app as a plugin. This react app
renders automatically its main component when the worpress page is
loaded. It is unnecesary.

Implementing lazy-loading fixes this issue. When the final user scroll
down to the react-app area the component is leazy loaded. Otherwise
it is neither loaded not rendered.